Fixed to check for Extra edition many times but not to publish repeated ones
Now the config update does not skip the day after getting the extra edition.
It keeps checking for new editions (since they can appear multiple times).
To avoid repeating the extra editions, it checks for a letter in the edition
(e.g. '132-A') and skips the articles with that letter (or lower).

We also included filters and config files for testing, that send the results to
the channel 'sistemas'.