Fixed issues regarding remote operation, next day data change, documentation and debug messages
This version should be working properly.

TODO: currently the downloaded URL is saved to the captured list even if the posting failed.
This should be changed in the future. The difficculty is that the posting happens every 20
articles, so the registering in the list must occur in a similar fashion.